scripts/hand-cursor/requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.26.0
pyobjc-framework-Quartz>=10.0
pyyaml>=6.0
